Update sequelize.js
This is a proposal which would support dynamic model options. For example, if you would like to re-use a model but change the table name for database performance reasons. it would be easy to inject this using an option while importing the model. I'm not sure exactly how the argument length block was used considering the JsDoc - definitely something to look at before merging.